+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -1385,6 +1385,7 @@ class ConfigurationTest extends TestCase
             ],
             'scheduler' => [
                 'enabled' => !class_exists(FullStack::class) && class_exists(SchedulerTransportFactory::class),
+                'use_messenger_routing' => null,
             ],
             'exceptions' => [],
             'webhook' => [