+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/scheduler_use_messenger_routing_legacy.php
@@ -1,0 +1,6 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'scheduler' => true,
+    'messenger' => true,
+]);